Module_06/video-course-app/src/app/components/header/header-user/header-user.component.ts
import { Component, OnInit } from '@angular/core';
import { User } from '../../../models/user';
import { AuthService } from '../../../services/auth.service';
import { Router } from '@angular/router';

@Component({
  selector: 'app-header-user',
  templateUrl: './header-user.component.html',
  styleUrls: ['./header-user.component.css']
})
export class HeaderUserComponent implements OnInit {

  user: User;

  constructor(private authService: AuthService, private router: Router) { }

  ngOnInit(): void {
    this.user = this.authService.getUserInfo();
  }

  onUserLinkClick(): void {

  }

  onLogoffLinkClick(): void {
    this.authService.logout();
    this.router.navigate(['/login']);
  }
}
